added odometer into constants

+++ add-transaction-to-fleetio/src/main/java/gcfv2/fleetio/Meter_entry_attributes.java
@@ -1,18 +1,18 @@
 package gcfv2.fleetio;
 
 public class Meter_entry_attributes {
-	int value;
+	double value;
 
-	public Meter_entry_attributes(int value) {
+	public Meter_entry_attributes(double value) {
 		super();
 		this.value = value;
 	}
 
-	public int getValue() {
+	public double getValue() {
 		return value;
 	}
 
-	public void setValue(int value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
